MSD radix sort
This is a preliminary radix sort implementation that avoids recursive
function calls with some extremely idiomatic Python; hopefully the comments
will make its functionality clear to those not well versed in itertools
trickery.  However, I'm going to rewrite this in the standard recursive way,
because this implementation has a glaring flaw: it always iterates over
every subpartition -- even those that cannot be partitioned further -- so the
running time is always the worst possible proportional to the length
of the longest string in the set.  I suppose another set of "open partitions"
could be maintained, but that's getting just too convoluted for my tastes.